requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via
    #   django
    #   uvicorn
azure-core==1.20.1
    # via azure-identity
azure-identity==1.7.1
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via uvicorn
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==35.0.0
    # via
    #   azure-identity
    #   msal
    #   pyjwt
django==3.2.9
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-yasg
django-cors-headers==3.10.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   drf-yasg
djangorestframework-gis==0.17
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.0.1
    # via jinja2
msal==1.16.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==0.3.0
    # via azure-identity
packaging==21.2
    # via drf-yasg
portalocker==1.7.1
    # via msal-extensions
psycopg2-binary==2.9.1
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via msal
pyparsing==2.4.7
    # via packaging
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   azure-core
    #   coreapi
    #   msal
ruamel.yaml==0.17.17
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sentry-sdk==1.4.3
    # via -r requirements.in
six==1.16.0
    # via
    #   azure-core
    #   azure-identity
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.15.0
    # via -r requirements.in
werkzeug==2.0.2
    # via -r requirements.in
whitenoise==5.3.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
